Minor improvements in Scheduler (#4785)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR introduces a few minor improvements in `Scheduler` class before
a great refactor. Feel free to review commits one-by-one.

## Test plan

1. BokehExample + <kbd>r</kbd>
2. runOnUI / runOnJS example
